+++ src/main/kotlin/no/nav/familie/pdf/pdf/visningsvarianter/HåndterVisningsvariant.kt
@@ -1,10 +1,10 @@
 package no.nav.familie.pdf.pdf.visningsvarianter
 
 import com.itextpdf.layout.element.Div
-import no.nav.familie.pdf.pdf.Innholdsfortegnelse.hentOversettelse
-import no.nav.familie.pdf.pdf.Innholdsfortegnelse.håndterRekursivVerdiliste
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 import no.nav.familie.pdf.pdf.domain.VisningsVariant
+import no.nav.familie.pdf.pdf.hentOversettelse
+import no.nav.familie.pdf.pdf.pdfElementer.håndterRekursivVerdiliste
 import no.nav.familie.pdf.pdf.pdfElementer.lagOverskriftH4
 import no.nav.familie.pdf.pdf.pdfElementer.lagTekstElement
