+++ src/test/kotlin/no/nav/familie/pdf/pdf/InnholdsfortegnelseTest.kt
@@ -1,8 +1,8 @@
 package no.nav.familie.pdf.pdf
 
 import com.itextpdf.layout.element.Div
-import no.nav.familie.pdf.pdf.Innholdsfortegnelse.håndterRekursivVerdiliste
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
+import no.nav.familie.pdf.pdf.pdfElementer.håndterRekursivVerdiliste
 import org.junit.jupiter.api.Test
 import kotlin.test.assertTrue
